requirements.txt
Django==5.2.8
psycopg2-binary==2.9.9
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
django-ratelimit==4.1.0
django-filter==23.5
Pillow==10.1.0
django-storages==1.14.2
django-compressor==4.4
django-debug-toolbar==4.2.0
django-axes==6.2.0
django-health-check==3.17.0
sentry-sdk[django]==1.38.0
python-decouple==3.8
django-redis==5.4.0
django-ckeditor==6.7.0
django-simple-captcha==0.6.0
django-recaptcha==3.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-environ==0.11.2
PyJWT==2.8.0
psutil==5.9.6
uvicorn==0.24.0
prometheus-client==0.19.0
structlog==23.2.0
django-silk==5.0.4
django-cachalot==2.6.1
django-celery-beat==2.5.0
flower==2.0.1
daphne==4.0.0